requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
cryptography==3.4.8
fastapi==0.115.2
greenlet==1.1.2
importlib-metadata==4.6.4
more-itertools==8.10.0
oauthlib==3.2.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.3.0
pynvim==0.4.2
pyparsing==2.4.7
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==5.4.1
starlette==0.39.2
typing_extensions==4.12.2
uvicorn==0.31.1
redis==5.0.0
passlib==1.7.4